Finally got dev setup working consistently
