+++ webapp/Dockerfile
@@ -1,12 +1,10 @@
-FROM python:3.6-alpine
+FROM python:3.6-onbuild
 
 ADD https://github.com/ufoscout/docker-compose-wait/releases/download/2.0.0/wait /wait
 RUN chmod +x /wait
 
-RUN mkdir /webapp
-WORKDIR /webapp
-VOLUME /webapp
-COPY requirements.txt /webapp/
-RUN pip install -r requirements.txt
-COPY . /webapp/
-CMD /wait && python app.py
+# CMD /wait && python app.py
+
+# CMD /wait && python -m flask run --host=0.0.0.0
+
+CMD /wait && gunicorn --config=gunicorn.py app:app
