+++ webapp/app/models.py
@@ -1,7 +1,7 @@
 from app import db
 from sqlalchemy.orm import relationship
 
-db.Model.metadata.reflect(db.engine)#change to (db.engine)
+db.Model.metadata.reflect(db.engine) #change to (db.engine)
 
 class Patient(db.Model):
     __table__ = db.Model.metadata.tables['patients']
