+++ webapp/gunicorn.py
@@ -1,0 +1,7 @@
+# gunicorn.py
+import os
+
+if os.environ.get('MODE') == 'dev':
+    reload = True
+
+bind = '0.0.0.0:5000'
