+++ webapp/requirements.txt
@@ -1,4 +1,5 @@
 flask
+gunicorn
 PyMySQL
 flask-sqlalchemy
 flask-script
