--- webapp/run.py
@@ -1,5 +1,0 @@
-from app import app
-
-app.debug = True
-if __name__ == "__main__":
-    app.run()
