--- webapp/setup.py
@@ -1,10 +1,0 @@
-from setuptools import setup
-
-setup(
-    name='mpower_api',
-    packages=['mpower_api'],
-    include_package_data=True,
-    install_requires=[
-        'flask',
-    ],
-)
